+++ src/main/java/Audio/AudioConnection.java
@@ -6,9 +6,9 @@
 package Audio;
 
 import Resource.Emoji;
-import Command.*;
 import Main.*;
 import Utility.SmartLogger;
+import com.sedmelluq.discord.lavaplayer.track.AudioTrackState;
 
 import net.dv8tion.jda.core.entities.VoiceChannel;
 import net.dv8tion.jda.core.managers.AudioManager;
@@ -24,12 +24,27 @@ public class AudioConnection {
     private static VoiceChannel vc;
     private static AudioManager am;
     
+    /**
+     * Connect to a Voice Channel
+     * @param e
+     * @param inform
+     */
     public static void connect(MessageReceivedEvent e, boolean inform)
     {
         try {
+            //If the message is from Private Channel, return
             if(e.getChannelType() == e.getChannelType().PRIVATE)
                 return;
             
+            //Prevent user from requesting to join another voice channel
+            if(e.getGuild().getSelfMember().getVoiceState().getChannel() != e.getMember().getVoiceState().getChannel() //When the user and the bot are in a different channel
+                    && Main.guilds.get(e.getGuild().getId()).getPlayer().getPlayingTrack() != null //When the player is playing
+                    && !Main.guilds.get(e.getGuild().getId()).getPlayer().isPaused() ) { //When the player is not paused
+                e.getChannel().sendMessage(Emoji.error + " I am already playing songs in a voice channel.").queue();
+                return;
+            }
+            
+            //Set Voice Channels and Text Channels
             Main.guilds.get(e.getGuild().getId()).setVc(e.getMember().getVoiceState().getChannel());
             Main.guilds.get(e.getGuild().getId()).setTc(e.getTextChannel());
             
@@ -53,13 +68,25 @@ public class AudioConnection {
     public static void disconnect(MessageReceivedEvent e, boolean inform)
     {
         try {
+            //Inform user that the bot is not in a Voice Channel
+            if(e.getGuild().getSelfMember().getVoiceState().getChannel() == null) {
+                e.getChannel().sendMessage(Emoji.error + " I am not in a voice channel.").queue();
+                return;
+            }
+            
+            //Prevent user from commanding the bot too leave a voice channel
+            if(e.getGuild().getSelfMember().getVoiceState().getChannel() != e.getMember().getVoiceState().getChannel()) {
+                e.getChannel().sendMessage(Emoji.error + " Do not tell me to leave a voice channel when you are not in it.").queue();
+                return;
+            }
+            
             am = e.getGuild().getAudioManager();
             am.closeAudioConnection();
         
         //Inform the users that the bot joined a voice channel
         if(inform)
             e.getChannel().sendMessage(Emoji.globe + " Left Voice Channel `" + Main.guilds.get(e.getGuild().getId()).getVc().getName() + "`").queue();
-        } catch (NullPointerException npe) {
+        } catch (NullPointerException npe) { 
             e.getChannel().sendMessage(Emoji.error + " I am not in a voice channel.").queue();
         }
     }